api-gateway/requirements.txt
python-consul===1.1.0
fastapi===0.111.0
requests===2.31.0
uvicorn===0.29.0
pyasyncore===1.0.4
hazelcast-python-client===5.3.0
fastapi_utils===0.6.0
typing-inspect===0.9.0
httpx===0.27.0